fs: fix compilation on non-Linux/macOS/Windows
