+++ internal/fs/fs_local_nofd.go
@@ -1,0 +1,7 @@
+//go:build !linux && !darwin && !windows
+
+package fs
+
+func newFdLocalFile(name string, flag int, metadataOnly bool) (File, error) {
+	panic("not supported")
+}
